Move sprite atlas as part of the scene
