+++ core/src/scene/spriteAtlas.cpp
