+++ core/src/scene/spriteAtlas.h
